Move storage_sync modules under tenant
