+++ pageserver/src/tenant/storage_sync/delete.rs
